fix(tests): update expected size in BuildRegistrationSignBytes test
Adjusted the expected size assertion in the RegistrationSignPayloadTest to reflect the correct calculation based on the new hybrid KEM public key size. This change ensures the test accurately verifies the output of the BuildRegistrationSignBytes function.

+++ src/feature/chat/ChatController.cpp
@@ -350,7 +350,7 @@ void ChatController::InsertEmojiIntoDraft(const std::string& emoji, bool restore
   }
   draft->SetValue(next.c_str());
   const Rml::String value = draft->GetValue();
-  const int end = Rml::StringUtilities::ConvertByteOffsetToCharacterOffset(value, value.size());
+  const int end = Rml::StringUtilities::ConvertByteOffsetToCharacterOffset(value, static_cast<int>(value.size()));
   if (!restore_composer_focus) {
     // Keyboard-panel multi-insert: caret only (unfocused SetSelectionRange does not OSK).
     draft->SetSelectionRange(end, end);
@@ -1862,7 +1862,7 @@ void ChatController::OnShellLayoutSynced() {
   }
   draft->Focus();
   const Rml::String value = draft->GetValue();
-  const int end = Rml::StringUtilities::ConvertByteOffsetToCharacterOffset(value, value.size());
+  const int end = Rml::StringUtilities::ConvertByteOffsetToCharacterOffset(value, static_cast<int>(value.size()));
   draft->SetSelectionRange(end, end);
 }
